feat: add new enable-integration-tests profile
Integration tests generally require credentials and access to external
GCP services which is not always available to every contributor.

This change skips the running of integration tests by default (they are
still compiled) in favor of explicitly enabling. To enable the running
of integration tests activate the `enable-integration-tests` profile
when executing maven (command line are: `-Penable-integration-tests`)